Add logic to update search index, move type to module.
